This commits adds a BestContructorSelector to prevent generation of faulty constructor in the built() method. This addresses missorted parameters case (issue #35 and also https://plugins.jetbrains.com/plugin/6585-builder-generator#comment=18490) and other tricky cases like multiple constructors, or when the constructor parameter list does not match the selected fields.
Description :
First, the BestContructorSelector algorithm looks for a constructor that matches exactly the fields to be set by constructor.
If the search fails, it looks for the constructor that have all needed fields plus extra ones, and the lowest parameter count.
If the search fails, it looks for the constructor, that have the most of needed fields, and the lowest parameter count.
If no constructor is found (null returned), it means that the default constructor is the best constructor.

If the best constructor found has extra parameters, a specific method fills up the extra parameters with default values according to the type.

Other improvements and bugfixes:
  - When using Use Single Field, the source class must have zero args constructor available. If no a specific message is displayed and the creation is refused.
  - If no fields are set by setters or assignments in the build() method, the constructor is output inline with the return statement (issue #26).

Code cleanup:
Replaced usage of org.junit.Assert I introduced in my previous commits, by org.assertj.core.api.Assertions for the sake of consistency.

Checkstyle:
  - Added to suppressions.xml MagicNumberCheck for all test files
  - Added to suppressions.xml all checks for MemberChooser.java and removed checkstyle occurrences in MemberChooser.java itself
  - The checkstyle plugin in build.gradle can be restored, as the build is now OK with the previous settings